third_party/blink/web_tests/wpt_internal/attribution-reporting/simple-event-level-report-test.sub.https.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
promise_test(async t => {
  const interval = 100;
  const sourceEvent = {
    source_event_id: '9153495207648170476',
    destination: `https://{{host}}`,
  };
  const trigger = {event_trigger_data: [{'trigger_data': '2'}]};
  await resetEventLevelReports();
  registerAttributionSrc('Attribution-Reporting-Register-Source', sourceEvent);
  // Adds delay to reduce chance of race condition between source registration and
  // trigger registration.
  await delay(interval);
  registerAttributionSrc(
      'Attribution-Reporting-Register-Trigger', trigger);
  const payload = await pollEventLevelReports(interval);
  assert_equals(payload.reports.length, 1);
  const report = JSON.parse(payload.reports[0]);
  // The trigger data is sanitized to "0" because event sources are limited to 1
  // bit.
  assert_equals(report.trigger_data, '0');
  assert_equals(report.source_event_id, sourceEvent.source_event_id);
  assert_equals(report.source_type, 'event');
  assert_equals(report.attribution_destination, sourceEvent.destination);
  assert_greater_than(report.randomized_trigger_rate, 0.0);
  assert_less_than(report.randomized_trigger_rate, 1.0);
  assert_own_property(report, 'report_id');
  assert_equals(typeof report.report_id, 'string');
  assert_not_own_property(report, 'source_debug_key');
  assert_not_own_property(report, 'trigger_debug_key');
}, 'Ensure attribution report is received.');
</script>
